msm: kgsl: Fix memory entry leak when calling adreno_convertaddr
When adreno_convertaddr is called the memory searched in this function
may have a memory entry associated with it if this allocation belongs
to user space. A reference to this memory entry is taken when
this function is called which should be returned when the caller is
done using the memory. Made changes to return a reference to the
entry when calling this function which is returned when caller is
done using the memory.

Change-Id: I610553fd4b119d3c47cbfcec970d50046dbb2863
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Lynus Vaz <lvaz@codeaurora.org>

(cherry picked from commit 58ee86d4b29fc8d726a7fef3a7a410de70f6c9b2)

(cherry picked from commit 58ee86d4b29fc8d726a7fef3a7a410de70f6c9b2)