ASoC: msm: qdsp6v2: check proxy port is open when closing
The current logic opens the afe proxy port only if the
proxy instance virtual port is opened. If only
pcm instance virtual port is opened and a close of the
afe proxy port is attempted, this causes the device to
crash due to a wait event timeout as the driver tries
to close an unopened afe port.
The change ensures that only if both the virtual ports
are opened, which in effect starts the proxy port,
the afe close commands for the proxy port is sent to DSP,
else return immediately.

Change-Id: I9450001c608208f3d8da22b2881a88db1c21d950
Signed-off-by: Aravind Kumar <akumark@codeaurora.org>

(cherry picked from commit fff4305a70962763da8b52582ca1dfa698794bd1)

(cherry picked from commit fff4305a70962763da8b52582ca1dfa698794bd1)